+++ app.py
@@ -12,7 +12,8 @@ inyectar_css()
 
 pagina = st.navigation(
     [
-        st.Page("views/build_your_menu.py", title="Build your menu", icon="🥗"),
+        st.Page("views/menu_del_dia.py", title="Menú del día", icon="🥗"),
+        st.Page("views/menu_semanal.py", title="Menú semanal", icon="🗓️"),
         st.Page("views/editor_recetas.py", title="Editor de recetas", icon="🧑‍🍳"),
         st.Page("views/editor_ingredientes.py", title="Editor de ingredientes", icon="🥕"),
         st.Page("views/personas.py", title="Personas", icon="🧑‍🤝‍🧑"),